LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/navigate-replace-same-document-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: null is not an object (evaluating 'navigation.currentEntry.key')

PASS navigate() with history: 'replace' option
